Split predicate into submodules by factory
